Fix HTMLImageElement#currentSrc
The currentSrc property of HTMLImageElement is supposed to be a full
URL, not the possibly-relatife URI from the src attribute. This causes
the "async src complete test" test in
test/web-platform-tests/tests/html/semantics/embedded-content/the-img-element/img.complete.html
to fail. Make currentSrc an absolute URL, partially fixing the test.

Additionally, the currentSrc property is supposed to be updated even if
loading the image fails. Fix the updating of currentSrc to fix this case
too.